Renamed main directory to "com.bringframework"
