--- src/main/java/com/bring/library/App.java
@@ -1,6 +1,0 @@
-package com.bring.library;
-
-public class App {
-    public static void main(String[] args) {
-    }
-}
