--- src/main/java/com/bring/library/configuration/AppConfiguration.java
@@ -1,4 +1,0 @@
-package com.bring.library.configuration;
-
-public class AppConfiguration {
-}
