--- src/main/java/com/bring/library/exception/BoboNotFoundException.java
@@ -1,4 +1,0 @@
-package com.bring.library.exception;
-
-public class BoboNotFoundException {
-}
