--- src/main/java/com/bring/library/exception/EmptyClassLoaderException.java
@@ -1,7 +1,0 @@
-package com.bring.library.exception;
-
-public class EmptyClassLoaderException extends RuntimeException {
-    public EmptyClassLoaderException(String message) {
-        super(message);
-    }
-}
