--- src/main/java/com/bring/library/exception/EmptyDirectoryException.java
@@ -1,7 +1,0 @@
-package com.bring.library.exception;
-
-public class EmptyDirectoryException extends RuntimeException{
-    public EmptyDirectoryException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
