--- src/main/java/com/bring/library/exception/NotFoundAnnotationException.java
@@ -1,7 +1,0 @@
-package com.bring.library.exception;
-
-public class NotFoundAnnotationException extends RuntimeException {
-    public NotFoundAnnotationException(String message) {
-        super(message);
-    }
-}
