--- src/main/java/com/bring/library/registry/ApplicationRegistry.java
@@ -1,4 +1,0 @@
-package com.bring.library.registry;
-
-public class ApplicationRegistry {
-}
