--- src/main/java/com/bring/library/service/creator/BoboCreatorService.java
@@ -1,4 +1,0 @@
-package com.bring.library.service.creator;
-
-public interface BoboCreatorService {
-}
