--- src/main/java/com/bring/library/service/creator/impl/BoboCreatorServiceImpl.java
@@ -1,6 +1,0 @@
-package com.bring.library.service.creator.impl;
-
-import com.bring.library.service.creator.BoboCreatorService;
-
-public class BoboCreatorServiceImpl implements BoboCreatorService {
-}
