--- src/main/java/com/bring/library/service/processor/AnnotationProcessorService.java
@@ -1,4 +1,0 @@
-package com.bring.library.service.processor;
-
-public interface AnnotationProcessorService {
-}
