+++ src/main/java/com/bringframework/App.java
@@ -1,0 +1,6 @@
+package com.bringframework;
+
+public class App {
+    public static void main(String[] args) {
+    }
+}
