+++ src/main/java/com/bringframework/annotation/Configuration.java
@@ -1,0 +1,4 @@
+package com.bringframework.annotation;
+
+public @interface Configuration {
+}
