+++ src/main/java/com/bringframework/annotation/Dao.java
@@ -1,0 +1,11 @@
+package com.bringframework.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface Dao {
+}
