+++ src/main/java/com/bringframework/configuration/AppConfiguration.java
@@ -1,0 +1,4 @@
+package com.bringframework.configuration;
+
+public class AppConfiguration {
+}
