+++ src/main/java/com/bringframework/exception/BoboNotFoundException.java
@@ -1,0 +1,4 @@
+package com.bringframework.exception;
+
+public class BoboNotFoundException {
+}
