+++ src/main/java/com/bringframework/exception/ClassHasMoreOneInterfacesException.java
@@ -1,0 +1,7 @@
+package com.bringframework.exception;
+
+public class ClassHasMoreOneInterfacesException extends RuntimeException{
+    public ClassHasMoreOneInterfacesException(String message) {
+        super(message);
+    }
+}
