+++ src/main/java/com/bringframework/exception/CreateInstanceException.java
@@ -1,0 +1,7 @@
+package com.bringframework.exception;
+
+public class CreateInstanceException extends RuntimeException {
+    public CreateInstanceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
