+++ src/main/java/com/bringframework/exception/EmptyClassLoaderException.java
@@ -1,0 +1,7 @@
+package com.bringframework.exception;
+
+public class EmptyClassLoaderException extends RuntimeException {
+    public EmptyClassLoaderException(String message) {
+        super(message);
+    }
+}
