+++ src/main/java/com/bringframework/exception/EmptyDirectoryException.java
@@ -1,0 +1,7 @@
+package com.bringframework.exception;
+
+public class EmptyDirectoryException extends RuntimeException{
+    public EmptyDirectoryException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
