+++ src/main/java/com/bringframework/exception/InvalidFileNameException.java
@@ -1,0 +1,7 @@
+package com.bringframework.exception;
+
+public class InvalidFileNameException extends RuntimeException {
+    public InvalidFileNameException(String message) {
+        super(message);
+    }
+}
