+++ src/main/java/com/bringframework/exception/NotFoundAnnotationException.java
@@ -1,0 +1,7 @@
+package com.bringframework.exception;
+
+public class NotFoundAnnotationException extends RuntimeException {
+    public NotFoundAnnotationException(String message) {
+        super(message);
+    }
+}
