+++ src/main/java/com/bringframework/exception/SetValueToFieldException.java
@@ -1,0 +1,7 @@
+package com.bringframework.exception;
+
+public class SetValueToFieldException extends RuntimeException {
+    public SetValueToFieldException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
