+++ src/main/java/com/bringframework/registry/ApplicationRegistry.java
@@ -1,0 +1,4 @@
+package com.bringframework.registry;
+
+public class ApplicationRegistry {
+}
