+++ src/main/java/com/bringframework/service/creator/BoboCreatorService.java
@@ -1,0 +1,4 @@
+package com.bringframework.service.creator;
+
+public interface BoboCreatorService {
+}
