+++ src/main/java/com/bringframework/service/creator/impl/BoboCreatorServiceImpl.java
@@ -1,0 +1,6 @@
+package com.bringframework.service.creator.impl;
+
+import com.bringframework.service.creator.BoboCreatorService;
+
+public class BoboCreatorServiceImpl implements BoboCreatorService {
+}
