+++ src/main/java/com/bringframework/service/processor/AnnotationProcessorService.java
@@ -1,0 +1,4 @@
+package com.bringframework.service.processor;
+
+public interface AnnotationProcessorService {
+}
