+++ src/main/java/com/bringframework/service/processor/impl/AnnotationProcessorServiceImpl.java
@@ -1,0 +1,6 @@
+package com.bringframework.service.processor.impl;
+
+import com.bringframework.service.processor.AnnotationProcessorService;
+
+public class AnnotationProcessorServiceImpl implements AnnotationProcessorService {
+}
