+++ src/main/java/com/bringframework/util/AnnotationProcessorUtil.java
@@ -1,0 +1,4 @@
+package com.bringframework.util;
+
+public class AnnotationProcessorUtil {
+}
