+++ src/main/java/demonstration/project/service/impl/ServiceClass.java
@@ -1,7 +1,7 @@
 package demonstration.project.service.impl;
 
-import com.bring.library.annotation.Autowired;
-import com.bring.library.annotation.Service;
+import com.bringframework.annotation.Autowired;
+import com.bringframework.annotation.Service;
 import demonstration.project.dao.DaoInterface;
 import demonstration.project.service.ServiceInterface;
